on read error, GC and retry
On any read error, do a GC and retry. Read can fail due to
out-of-file-descriptors, so this gives us a good chance of preventing
that.

It would be better to trigger the GC from inside vips and just on open
fail, but this is mich harder.

We can't really do this for write, unfortunately, because write can take
a long time and we really don't want to retry that. read is quick and
harmless, so doing it twice shouldn't matter.