try to patch around mutate + libvips 8.9.1
libvips 8.9.1 had a terrible misfeature which would block metadata
changes unless the object had a ref_count of 1. MutableImage will
always have a ref_count of at least 2 (the parent gobject keeps a ref,
and we keep a ref to the copy ready to return to our caller), so we must
temporarily drop the refs to 1 around metadata changes.

See https://github.com/libvips/ruby-vips/issues/291