Reset concurrency to the initial default by setting 0 or nil
`Vips.concurrency_set(0)` suggests it'll set concurrency back to the
default value, but it does nothing if concurrency is already set, which
is always true because Vips.init has already configured thread pools.

The result is that if you set concurrency to any value, setting it to
zero will leave that value in place.

This reads the default concurrency immediately after Vips.init so we
can make good on having `concurrency_set 0` reset to default. Also
allows setting `nil` to reset to default.

Note that setting concurrency does not reconfigure existing thread
pools. It affects thread pools spawned for future operations.